Re-add WorldGroupManager as interface and AbstractWorldGroupManager
